refactor: separate StoryOperations::DestroyAll into its own file
